Exercice mini_faq selects 7/8

+++ 03-Databases/03-mini_faq/01-mini_faq-creationbdd.sql
@@ -26,7 +26,7 @@ CREATE TABLE categories
 (
 	category_name VARCHAR(30) PRIMARY KEY ,
     category_description VARCHAR(255),
-    category_order INT(3) NOT NULL UNIQUE
+    category_order TINYINT NOT NULL UNIQUE
 );
 
 CREATE TABLE categories_questions
